move google sheets to lib function
